docs: align README, CHANGELOG and test comments with the single-button changes
